requirements.txt
<<<<<<< HEAD
pandas
numpy
scikit-learn
matplotlib
seaborn
sqlalchemy
snowflake-connector-python
python-dotenv
=======
pandas
numpy
scikit-learn
matplotlib
seaborn
sqlalchemy
snowflake-connector-python
python-dotenv
>>>>>>> 7ac0f97 (update)
